brain/pyproject.toml
[project]
name = "ideacapital-brain"
version = "0.1.0"
description = "IdeaCapital AI Agent - Invention analysis, patent structuring, and prior art search"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langchain>=0.1.0",
    "langchain-google-vertexai>=0.0.6",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-pubsub>=2.19.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-firestore>=2.14.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "pinecone-client>=3.0.0",
    "openai>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
